move file, fix tests
