+++ refiner/app/api/v1/configurations/router.py
@@ -3,7 +3,7 @@ from fastapi import APIRouter
 from .activation import router as activation_router
 from .base import router as base_router
 from .codesets import router as codesets_router
-from .custom_codes.base import router as custom_codes_router
+from .custom_codes import router as custom_codes_router
 from .exports import router as exports_router
 from .locking import router as locking_router
 from .sections import router as sections_router
